+++ packages/nostrify/NIP05.test.ts
@@ -1,10 +1,11 @@
-import { assertEquals } from '@std/assert';
+import { assertEquals, assertRejects } from '@std/assert';
 import { returnsNext, stub } from '@std/testing/mock';
 
 import { NIP05 } from './NIP05.ts';
-import nostrJson from '../../fixtures/nostr.json' with { type: 'json' };
 
 Deno.test('NIP05.lookup', async () => {
+  const { default: nostrJson } = await import('../../fixtures/nostr.json', { with: { type: 'json' } });
+
   const fetch = stub(
     globalThis,
     'fetch',
@@ -21,5 +22,46 @@ Deno.test('NIP05.lookup', async () => {
   };
 
   assertEquals(result, expected);
+  fetch.restore();
+});
+
+// https://github.com/nostrability/nostrability/issues/143#issuecomment-2565772246
+Deno.test('NIP05.lookup with invalid values but valid profile pointer', async () => {
+  const { default: nostrJson } = await import('../../fixtures/lncal.json', { with: { type: 'json' } });
+
+  const fetch = stub(
+    globalThis,
+    'fetch',
+    returnsNext([
+      Promise.resolve(new Response(JSON.stringify(nostrJson))),
+    ]),
+  );
+
+  const result = await NIP05.lookup('elsat@lncal.com', { fetch });
+
+  const expected = {
+    pubkey: '17538dc2a62769d09443f18c37cbe358fab5bbf981173542aa7c5ff171ed77c4',
+    relays: undefined,
+  };
+
+  assertEquals(result, expected);
+  fetch.restore();
+});
+
+Deno.test('NIP05.lookup with invalid document', () => {
+  const fetch = stub(
+    globalThis,
+    'fetch',
+    returnsNext([
+      Promise.resolve(new Response(JSON.stringify({ names: 'yolo' }))),
+      Promise.resolve(new Response(JSON.stringify({}))),
+      Promise.resolve(new Response(JSON.stringify([]))),
+    ]),
+  );
+
+  assertRejects(() => NIP05.lookup('alex@gleasonator.dev', { fetch }));
+  assertRejects(() => NIP05.lookup('alex@gleasonator.dev', { fetch }));
+  assertRejects(() => NIP05.lookup('alex@gleasonator.dev', { fetch }));
+
   fetch.restore();
 });